okay working much better now.
It's super ugly, hard to follow, but the program no longer slows to a
crawl! What happens now for a really long dep graph is it just runs for
ever. It's good cause it's no longer stuck copying huge ExpandedPackage
data, but it's  bad cause clearly there's a recursive dependency that
isn't being caught by this, and I don't really know how to figure that
out :thinking: